Uncommented out code in deleteCustomerById to catch exception if id provided is not in database

+++ customers.js
@@ -63,22 +63,19 @@ const updateCustomer = async (pool, id, { fullName, email, password }) => {
 }
 
 const deleteCustomerById = async (pool, id) => {
-    // const {rows: customers} = await pool.query(`
-    //     SELECT * FROM customers 
-    //     WHERE id = $1;
-    // `, [id])
+    const {rows: customers} = await pool.query(`
+        SELECT * FROM customers 
+        WHERE id = $1;
+    `, [id])
     
-    // if (customers.length === 0) {
-    //     throw new Error(`Could not delete row because id provided (${id}) does not exist in table.`)
-    // } 
+    if (customers.length === 0) {
+        throw new Error(`Could not delete row because id provided (${id}) does not exist in table.`)
+    } 
 
-    // const {rows: customer} = 
     await pool.query(`
         DELETE FROM customers
         WHERE id = $1;
     `, [id])
-    
-    // return customer;
 }
 
 module.exports = {
